Work in progress
 * converting facades to use DeltaSpike Data repositories
 * wired Product to invoice row (needs to be rethought, probably should not reference at entity level at all, but just have "fill from product" in UI level)

+++ src/main/java/org/example/backend/AbstractEntity.java
@@ -17,16 +17,16 @@ public abstract class AbstractEntity implements Serializable, Cloneable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     
     @Version
     int version;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    protected void setId(Integer id) {
+    protected void setId(Long id) {
         this.id = id;
     }
